Check security hash before accepting downloaded files.
Fixes #7 in the simplest possible way. The current code will
ignore any files with a different security hash from the one
given by the developer. When ignoring, it emits a message
warning of a potentially malicious content. It will also always
emit a message when a downloaded file passes the security hash
check.